fixed highlighting
